backend/requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Agent Frameworks
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.3
langchain-anthropic==0.1.1
langgraph==0.0.20

# Vector Database
qdrant-client==1.7.0
sentence-transformers==2.3.1
rank-bm25==0.2.2

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Caching & Queue
redis==5.0.1
celery==5.3.6

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiohttp==3.9.1
tenacity==8.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
slowapi==0.1.9
